+++ Tests/Opc.Ua.InformationModel.Tests/RedundancyModelTests.cs
@@ -76,7 +76,7 @@ namespace Opc.Ua.InformationModel.Tests
             Assert.That(StatusCode.IsGood(dv.StatusCode), Is.True,
                 "RedundancySupport should be readable.");
 
-            int value = (int)dv.WrappedValue.GetInt32();
+            int value = dv.WrappedValue.GetInt32();
             // RedundancySupport enum: None=0, Cold=1, Warm=2,
             // Hot=3, Transparent=4, HotAndMirrored=5
             Assert.That(value, Is.InRange(0, 5),
@@ -90,7 +90,7 @@ namespace Opc.Ua.InformationModel.Tests
                 VariableIds.Server_ServerRedundancy_RedundancySupport)
                 .ConfigureAwait(false);
 
-            int redundancySupport = (int)dv.WrappedValue.GetInt32();
+            int redundancySupport = dv.WrappedValue.GetInt32();
 
             // ServerUriArray is only mandatory when redundancy != None
             if (redundancySupport == 0)
@@ -181,7 +181,7 @@ namespace Opc.Ua.InformationModel.Tests
                 VariableIds.Server_ServerRedundancy_RedundancySupport)
                 .ConfigureAwait(false);
 
-            int redundancySupport = (int)dv.WrappedValue.GetInt32();
+            int redundancySupport = dv.WrappedValue.GetInt32();
             if (redundancySupport == 0)
             {
                 Assert.Ignore(
